requirements.txt
asgiref==3.7.2
Brotli==1.0.9
dj-database-url==2.0.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
gunicorn==21.2.0
importlib-metadata==6.8.0
Markdown==3.4.4
packaging==23.1
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
zipp==3.16.2
